backend/lambda/process/requirements.txt
google-genai
Pillow
